EntriesコントローラーがFatになっていたのでリファクタリングした - GoogleAPI関係の処理をGoogleApiActionsとしてconcernsにまとめた - toast表示用のロジックをApplicationControllerで共通化 - Entriesテーブルを操作するロジックをEntryモデルに分離 - リサイズ関係の処理をEntryResizerデコレーターに分離
