+++ app/controllers/application_controller.rb
@@ -9,23 +9,6 @@ class ApplicationController < ActionController::Base
 
   private
 
-  def ensure_valid_access_token!
-    return if session[:token_expires_at].nil? || !current_user.token_expired?(session[:token_expires_at])
-
-    result = current_user.request_token_refresh(session[:refresh_token])
-    if result.present?
-      store_access_token(result)
-    else
-      sign_out current_user
-      redirect_to root_path, alert: t('activerecord.errors.messages.failed_to_refresh_token')
-    end
-  end
-
-  def store_access_token(result)
-    session[:access_token] = result['access_token']
-    session[:token_expires_at] = Time.current + result['expires_in'].seconds
-  end
-
   def set_selected_contest
     contest_id = params[:contest_id] || params[:id]
     return unless user_signed_in? && contest_id.present?
@@ -41,5 +24,11 @@ class ApplicationController < ActionController::Base
                        else
                          []
                        end
+  end
+
+  def append_turbo_toast(type, message)
+    turbo_stream.append('toast',
+                        partial: 'shared/toast',
+                        locals: { toasts: [{ type:, message: }] })
   end
 end
