+++ app/controllers/concerns/google_api_actions.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+module GoogleApiActions
+  extend ActiveSupport::Concern
+
+  private
+
+  def ensure_valid_access_token!
+    return if session[:token_expires_at].nil? || !current_user.token_expired?(session[:token_expires_at])
+
+    result = current_user.request_token_refresh(session[:refresh_token])
+    if result.present?
+      store_access_token(result)
+    else
+      sign_out current_user
+      redirect_to root_path, alert: t('activerecord.errors.messages.failed_to_refresh_token')
+    end
+  end
+
+  def store_access_token(result)
+    session[:access_token] = result['access_token']
+    session[:token_expires_at] = Time.current + result['expires_in'].seconds
+  end
+
+  def set_drive_service
+    @drive_service = GoogleDriveService.new(session[:access_token])
+  end
+end
